Fixing NbValideContrat

+++ back/src/main/java/com/example/springproject/Service/ContratService.java
@@ -76,7 +76,7 @@ public class ContratService implements IContratService {
     }
     @Override
     public Integer nbContratsValides(Date endDate, Date startDate) {
-        return myRepository.countContratByDateDebutContratAfterAndDateFinContratBefore(endDate,startDate);
+        return myRepository.countContratByDateDebutContratAfterAndDateFinContratBefore(endDate,startDate).intValue();
     }
 
     @Override
